Change package of cache merge policies for backward compat, fixes #6439
