+++ hazelcast/src/test/java/com/hazelcast/cache/merge/CacheMergePolicyProviderTest.java
@@ -3,12 +3,7 @@ package com.hazelcast.cache.merge;
 import com.hazelcast.cache.BuiltInCacheMergePolicies;
 import com.hazelcast.cache.CacheMergePolicy;
 import com.hazelcast.cache.impl.merge.policy.CacheMergePolicyProvider;
-import com.hazelcast.cache.impl.merge.policy.HigherHitsCacheMergePolicy;
-import com.hazelcast.cache.impl.merge.policy.LatestAccessCacheMergePolicy;
-import com.hazelcast.cache.impl.merge.policy.PassThroughCacheMergePolicy;
-import com.hazelcast.cache.impl.merge.policy.PutIfAbsentCacheMergePolicy;
 import com.hazelcast.config.InvalidConfigurationException;
-import com.hazelcast.core.HazelcastException;
 import com.hazelcast.test.HazelcastParallelClassRunner;
 import com.hazelcast.test.HazelcastTestSupport;
 import com.hazelcast.test.annotation.ParallelTest;
