+++ hazelcast/src/test/java/com/hazelcast/cache/merge/CacheSplitBrainTest.java
@@ -3,10 +3,6 @@ package com.hazelcast.cache.merge;
 import com.hazelcast.cache.CacheEntryView;
 import com.hazelcast.cache.CacheMergePolicy;
 import com.hazelcast.cache.impl.HazelcastServerCachingProvider;
-import com.hazelcast.cache.impl.merge.policy.HigherHitsCacheMergePolicy;
-import com.hazelcast.cache.impl.merge.policy.LatestAccessCacheMergePolicy;
-import com.hazelcast.cache.impl.merge.policy.PassThroughCacheMergePolicy;
-import com.hazelcast.cache.impl.merge.policy.PutIfAbsentCacheMergePolicy;
 import com.hazelcast.config.CacheConfig;
 import com.hazelcast.config.Config;
 import com.hazelcast.core.Hazelcast;
