+++ hazelcast/src/test/java/com/hazelcast/cache/merge/HigherHitsCacheMergePolicyTest.java
@@ -1,7 +1,6 @@
 package com.hazelcast.cache.merge;
 
 import com.hazelcast.cache.CacheMergePolicy;
-import com.hazelcast.cache.impl.merge.policy.HigherHitsCacheMergePolicy;
 import com.hazelcast.test.HazelcastParallelClassRunner;
 import com.hazelcast.test.annotation.ParallelTest;
 import com.hazelcast.test.annotation.QuickTest;
